+++ packages/Webkul/Admin/src/Resources/views/customers/groups/index.blade.php
@@ -32,9 +32,13 @@
 
                 <!-- DataGrid -->
                 <x-admin::datagrid src="{{ route('admin.customers.groups.index') }}" ref="datagrid">
+                    @php
+                        $hasPermission = bouncer()->hasPermission('customers.groups.edit') || bouncer()->hasPermission('customers.groups.delete');
+                    @endphp
+
                     <!-- DataGrid Header -->
                     <template #header="{ columns, records, sortPage, applied}">
-                        <div class="row grid grid-cols-4 grid-rows-1 gap-[10px] items-center px-[16px] py-[10px] border-b-[1px] border-gray-300 text-gray-600 bg-gray-50 font-semibold">
+                        <div class="row grid grid-cols-{{ $hasPermission ? '4' : '3' }} grid-rows-1 gap-[10px] items-center px-[16px] py-[10px] border-b-[1px] border-gray-300 text-gray-600 bg-gray-50 font-semibold">
                             <div
                                 class="flex gap-[10px] cursor-pointer"
                                 v-for="(columnGroup, index) in ['id', 'code', 'name']"
@@ -65,9 +69,11 @@
                             </div>
     
                             <!-- Actions -->
-                            <p class="flex gap-[10px] justify-end">
-                                @lang('admin::app.components.datagrid.table.actions')
-                            </p>
+                            @if ($hasPermission)
+                                <p class="flex gap-[10px] justify-end">
+                                    @lang('admin::app.components.datagrid.table.actions')
+                                </p>
+                            @endif
                         </div>
                     </template>
 
@@ -76,7 +82,7 @@
                         <div
                             v-for="record in records"
                             class="row grid gap-[10px] items-center px-[16px] py-[16px] border-b-[1px] border-gray-300 text-gray-600 transition-all hover:bg-gray-50"
-                            style="grid-template-columns: repeat(4, 1fr);"
+                            :style="'grid-template-columns: repeat(' + (record.actions.length ? 4 : 3) + ', 1fr);'"
                         >
                             <!-- Id -->
                             <p v-text="record.id"></p>
@@ -89,23 +95,27 @@
 
                             <!-- Actions -->
                             <div class="flex justify-end">
-                                <a @click="id=1; editModal(record)">
-                                    <span
-                                        :class="record.actions['0'].icon"
-                                        class="cursor-pointer rounded-[6px] p-[6px] text-[24px] transition-all hover:bg-gray-200 max-sm:place-self-center"
-                                        :title="record.actions['0'].title"
-                                    >
-                                    </span>
-                                </a>
+                                @if (bouncer()->hasPermission('customers.groups.edit'))
+                                    <a @click="id=1; editModal(record)">
+                                        <span
+                                            :class="record.actions['0'].icon"
+                                            class="cursor-pointer rounded-[6px] p-[6px] text-[24px] transition-all hover:bg-gray-200 max-sm:place-self-center"
+                                            :title="record.actions['0'].title"
+                                        >
+                                        </span>
+                                    </a>
+                                @endif
 
-                                <a @click="deleteModal(record.actions['1']?.url)">
-                                    <span
-                                        :class="record.actions['1'].icon"
-                                        class="cursor-pointer rounded-[6px] p-[6px] text-[24px] transition-all hover:bg-gray-200 max-sm:place-self-center"
-                                        :title="record.actions['1'].title"
-                                    >
-                                    </span>
-                                </a>
+                                @if (bouncer()->hasPermission('customers.groups.delete'))
+                                    <a @click="deleteModal(record.actions['1']?.url)">
+                                        <span
+                                            :class="record.actions['1'].icon"
+                                            class="cursor-pointer rounded-[6px] p-[6px] text-[24px] transition-all hover:bg-gray-200 max-sm:place-self-center"
+                                            :title="record.actions['1'].title"
+                                        >
+                                        </span>
+                                    </a>
+                                @endif
                             </div>
                         </div>
                     </template>
